split invoiceItems out of invoices
