+++ typescript/src/shared/tools.ts
@@ -13,9 +13,9 @@ import {createPaymentLinkPrompt} from '@/shared/paymentLinks/prompts';
 import {
   createInvoicePrompt,
   listInvoicesPrompt,
-  createInvoiceItemPrompt,
   finalizeInvoicePrompt,
 } from '@/shared/invoices/prompts';
+import {createInvoiceItemPrompt} from '@/shared/invoiceItems/prompts';
 import {retrieveBalancePrompt} from '@/shared/balance/prompts';
 import {createRefundPrompt} from '@/shared/refunds/prompts';
 import {searchDocumentationPrompt} from '@/shared/documentation/prompts';
@@ -37,9 +37,9 @@ import {createPaymentLinkParameters} from '@/shared/paymentLinks/parameters';
 import {
   createInvoiceParameters,
   listInvoicesParameters,
-  createInvoiceItemParameters,
   finalizeInvoiceParameters,
-} from './invoices/parameters';
+} from '@/shared/invoices/parameters';
+import {createInvoiceItemParameters} from '@/shared/invoiceItems/parameters';
 import {retrieveBalanceParameters} from '@/shared/balance/parameters';
 import {createRefundParameters} from '@/shared/refunds/parameters';
 import {searchDocumentationParameters} from '@/shared/documentation/parameters';
