+++ typescript/src/test/shared/invoices/parameters.test.ts
@@ -1,7 +1,6 @@
 import {
   createInvoiceParameters,
   listInvoicesParameters,
-  createInvoiceItemParameters,
   finalizeInvoiceParameters,
 } from '@/shared/invoices/parameters';
 
@@ -38,24 +37,6 @@ describe('listInvoicesParameters', () => {
     const fields = Object.keys(parameters.shape);
     expect(fields).toEqual(['limit']);
     expect(fields.length).toBe(1);
-  });
-});
-
-describe('createInvoiceItemParameters', () => {
-  it('should return the correct parameters if no context', () => {
-    const parameters = createInvoiceItemParameters({});
-
-    const fields = Object.keys(parameters.shape);
-    expect(fields).toEqual(['customer', 'price', 'invoice']);
-    expect(fields.length).toBe(3);
-  });
-
-  it('should return the correct parameters if customer is specified', () => {
-    const parameters = createInvoiceItemParameters({customer: 'cus_123'});
-
-    const fields = Object.keys(parameters.shape);
-    expect(fields).toEqual(['price', 'invoice']);
-    expect(fields.length).toBe(2);
   });
 });
